requirements.txt
absl-py==2.0.0
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.0
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.4
fire==0.5.0
fonttools==4.42.1
fsspec==2023.9.2
ghostscript==0.7
google-auth==2.23.2
google-auth-oauthlib==1.0.0
grpcio==1.58.0
huggingface-hub==0.17.3
idna==3.4
img2table==1.2.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.12.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.4.1
jupyter-client==8.3.1
jupyter-core==5.3.2
kiwisolver==1.4.5
lit==17.0.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdftopng==0.2.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
polars==0.19.3
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycparser==2.21
Pygments==2.16.1
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
pyparsing==3.1.1
pypdf==3.16.2
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sahi==0.11.14
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.31.0
shapely==2.0.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
triton==2.0.0
typing-extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.43
ultralyticsplus==0.0.28
urllib3==2.0.5
wcwidth==0.2.7
werkzeug==2.3.7
XlsxWriter==3.1.5
zipp==3.17.0
